Save chat webhook messages
Add support to store/browse messages sent from Sendbird through a webhook.
https://docs.sendbird.com/platform/webhooks

First - let's implement a debugging piece of code, to record all the messages to figure out the ones we actually need.
After that - we'll filter the ones we need - we only care about when was the last move in the game, and the user messages.
- Store debugging messages
-  Store last game update data